third_party/WebKit/LayoutTests/bluetooth/server/getPrimaryService/disconnect-called-during-error.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  return setBluetoothFakeAdapter('HeartRateAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['heart_rate']}],
      optionalServices: ['battery_service']}))
    .then(device => device.gatt.connect())
    .then(gattServer => {
      let promise = assert_promise_rejects_with_message(
        gattServer.getPrimaryService('battery_service'),
        new DOMException('GATT Server is disconnected. ' +
                           'Cannot retrieve services. ' +
                           '(Re)connect first with `device.gatt.connect`.',
                         'NetworkError'));
      gattServer.disconnect();
      return promise;
    });
}, 'disconnect() called during a getPrimaryService call that fails. Reject ' +
   'with NetworkError.');
</script>
